EE-12559 Never run archiving in a non-production environment.
As we only scale down the non-prod pods every night, we do not know if
they will be up, and archiving will fail if they are not.  Therefore
we will never run archiving in non-prod automatically.

Archiving can be run in prod manually by either hitting the archiving
endpoint manually, or adjusting the cron schedule for the archving
cron job with kubectl.